Divide into two subfolders, submarine and ordinal
Vite is adding the ability to provide multiple entry points, but it's unclear
how this works with the react plugin. Instead of spending time on this now, I'm
copying the full vite-react single-page app to two subfolders, one for the submarine swap,
the other for the ordinal swap. For each new thin swap client, create a subfolder.
Once it's clear on how to provide multiple entry points, switch to having one
multipage app with shared code.